+++ docs/versionhistory.rst
@@ -44,6 +44,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - **BACKWARDS INCOMPATIBLE** ``Lock`` and ``Condition`` can now only be released by the task that
   acquired them. This behavior is now consistent on all backends whereas previously only Trio
   enforced this.
+- **BACKWARDS INCOMPATIBLE** The ``SocketStream`` and ``SocketListener`` classes were made
+  non-generic
 - Dropped Curio as a backend (see the :doc:`FAQ <faq>` as for why)
 - Added the ``run_sync_from_thread()`` function
 - Added the ``FileReadStream`` and ``FileWriteStream`` classes